requirements.txt
fastapi~=0.128.0
uvicorn~=0.20.0
pydantic
sqlalchemy
python-dotenv